Add level actors
Actors can be placed in levels, one per tile. Player will start at
"player start"

+++ src/editor_core.h
@@ -8,7 +8,7 @@ namespace Editor {
 		u8 chrPaletteIndex[2];
 		u8 chrSelection[2];
 
-		u32 levelClipboard[((VIEWPORT_WIDTH_TILES / 2) + 1) * ((VIEWPORT_HEIGHT_TILES / 2) + 1)];
+		u8 levelClipboard[((VIEWPORT_WIDTH_TILES / 2) + 1) * ((VIEWPORT_HEIGHT_TILES / 2) + 1)];
 		ImVec2 levelSelectionSize;
 		ImVec2 levelSelectionOffset;
